release: publish release artifact (#67)
GitHub's stability guarantee for the archive is iffy, and we want metrics on downloads. SeeGitHub's stability guarantee for the archive is iffy, and we want metrics on downloads. SeeGitHub's stability guarantee for the archive is iffy, and we want metrics on downloads. See https://github.com/bazel-contrib/SIG-rules-authors/issues/11#issuecomment-1411222578